Fixed crash when TownyChat was not installed
